refactor(users): extract RevokeSessionsDialog into its own component file
